# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Syrian Car Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cars:
    get:
      operationId: listCars
      tags: [cars]
      summary: List all cars with optional filters
      parameters:
        - name: brand
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: minYear
          in: query
          schema:
            type: integer
        - name: maxYear
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minMileage
          in: query
          schema:
            type: integer
        - name: maxMileage
          in: query
          schema:
            type: integer
        - name: province
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: fuelType
          in: query
          schema:
            type: string
        - name: transmission
          in: query
          schema:
            type: string
        - name: saleType
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of cars
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarListResponse"

    post:
      operationId: createCar
      tags: [cars]
      summary: Create a new car listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarBody"
      responses:
        "201":
          description: Car created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
...
  /cars/{id}:
    get:
      operationId: getCar
      tags: [cars]
      summary: Get car listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Car details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDetail"
        "404":
          description: Car not found

    patch:
      operationId: updateCar
      tags: [cars]
      summary: Update car listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCarBody"
      responses:
        "200":
          description: Updated car
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"

    delete:
      operationId: deleteCar
      tags: [cars]
      summary: Delete car listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Car deleted
  /cars/{id}/images:
    post:
      operationId: addCarImage
      tags: [cars]
      summary: Add image to car listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddImageBody"
      responses:
        "201":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarImage"
  /cars/{id}/promote:
    post:
      operationId: promoteCar
      tags: [cars]
      summary: Promote/feature a car listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Car promoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Car:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerPhoto:
          type: ["string", "null"]
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        mileage:
          type: integer
        fuelType:
          type: string
        transmission:
          type: string
        province:
          type: string
        city:
          type: string
        saleType:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isHighlighted:
          type: boolean
        isActive:
          type: boolean
        primaryImage:
          type: ["string", "null"]
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sellerId
        - sellerName
        - brand
        - model
        - year
        - price
        - mileage
        - fuelType
        - transmission
        - province
        - city
        - saleType
        - category
        - isFeatured
        - isHighlighted
        - isActive
        - viewCount
        - createdAt
    CarDetail:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerPhoto:
          type: ["string", "null"]
        sellerPhone:
          type: ["string", "null"]
        sellerRating:
          type: ["number", "null"]
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        mileage:
          type: integer
        fuelType:
          type: string
        transmission:
          type: string
        province:
          type: string
        city:
          type: string
        saleType:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isHighlighted:
          type: boolean
        isActive:
          type: boolean
        images:
          type: array
          items:
            $ref: "#/components/schemas/CarImage"
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sellerId
        - sellerName
        - brand
        - model
        - year
        - price
        - mileage
        - fuelType
        - transmission
        - province
        - city
        - saleType
        - category
        - isFeatured
        - isHighlighted
        - isActive
        - images
        - viewCount
        - createdAt
    CarImage:
      type: object
      properties:
        id:
          type: integer
        carId:
          type: integer
        imageUrl:
          type: string
        isPrimary:
          type: boolean
      required:
        - id
        - carId
        - imageUrl
        - isPrimary
    CarListResponse:
      type: object
      properties:
        cars:
          type: array
          items:
            $ref: "#/components/schemas/Car"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - cars
        - total
        - page
        - limit
    CreateCarBody:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        mileage:
          type: integer
        fuelType:
          type: string
        transmission:
          type: string
        province:
          type: string
        city:
          type: string
        saleType:
          type: string
        category:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
      required:
        - brand
        - model
        - year
        - price
        - mileage
        - fuelType
        - transmission
        - province
        - city
        - saleType
        - category
    UpdateCarBody:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        mileage:
          type: integer
        fuelType:
          type: string
        transmission:
          type: string
        province:
          type: string
        city:
          type: string
        saleType:
          type: string
        category:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    AddImageBody:
      type: object
      properties:
        imageUrl:
          type: string
        isPrimary:
          type: boolean
      required:
        - imageUrl
...